third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/xmlhttprequest-test-custom-headers-expected.txt

This test verifies that XMLHttpRequest properly handles list-valued headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.getResponseHeader("CONTENT-TYPE").indexOf("text/plain") is >= 0
PASS xhr.getResponseHeader("x-custom-header-single") is "single"
PASS xhr.getResponseHeader("x-custom-header-empty") is ""
PASS xhr.getResponseHeader("x-custom-header-list") is "one, two"
PASS successfullyParsed is true

TEST COMPLETE
